move adjust_author into separate ADAUTask of ReScanner
Extract adjust_author logic, get_author_mapping, and author dictionaries
from ReImporter into a new ADAUTask scanner task. The task only adjusts
authors for articles with KORREKTURSTAND=unvollständig.
