+++ service/ws_re/register/importer.py
@@ -5,10 +5,9 @@ from typing import Optional
 
 from pywikibot import Site, Page, Category
 
-from service.ws_re.register.authors import Authors
 from service.ws_re.register.lemma import Lemma
 from service.ws_re.register.registers import Registers
-from service.ws_re.template.article import Article
+from service.ws_re.scanner.tasks.adjust_author import adjust_author, get_author_mapping
 from tools import save_if_changed
 from tools.bots.cloud_bot import CloudBot
 
@@ -23,7 +22,7 @@ class ReImporter(CloudBot):
         super().__init__(wiki, debug, log_to_screen, log_to_wiki)
         self.registers = Registers(update_data=True)
         self.new_articles: dict[str, dict[str, str]] = {}
-        self.author_mapping = self.get_author_mapping()
+        self.author_mapping = get_author_mapping()
         self.tm_set = self.load_tm_set()
         self._create_neuland()
         self.current_year = datetime.now().year
@@ -65,7 +64,7 @@ class ReImporter(CloudBot):
                         article_text = self.get_text(article.volume.name, article.lemma)
                         if not article_text:
                             article_text = self.get_text_backup(article.volume.name, article)
-                        article_text = self.adjust_author(article_text, self.author_mapping)
+                        article_text = adjust_author(article_text, self.author_mapping)
                         article_text = self.adjust_end_column(article_text, register, idx)
                         article_text = article_text.replace("KORREKTURSTAND=Platzhalter",
                                                             "KORREKTURSTAND=unvollständig")
@@ -124,88 +123,6 @@ class ReImporter(CloudBot):
 [...]
 {{{{REAutor|{author}}}}}"""
         return parsed_article
-
-    ADDITIONAL_AUTHORS: dict[str, str] = {
-        "Franz Heinrich Weissbach": "Weissbach.",
-        "Hans von Arnim": "v. Arnim.",
-        "Paul Friedländer": "P. Friedländer.",
-        "Felix Jacoby": "F. Jacoby.",
-        "Hans von Geisau": "v. Geisau.",
-        "Paul Schoch-Bodmer": "Schoch.",
-        "Max Fluß": "Fluss.",
-        "August Burckhardt-Brandenberg": "Burckhardt.",
-        "Johannes Geffcken": "Geffcken.",
-        "Walther Eltester": "Eltester.",
-        "Fritz Geyer": "Geyer.",
-        "Bernhard große Kruse": "gr. Kruse.",
-        "Wilhelm Enßlin": "W. Enßlin.",
-        "Walter Friedrich Otto": "W. F. Otto.",
-        "Hans Philipp": "Philipp.",
-        "Schmidt": "Johanna Schmidt.",
-        "Albert William Van Buren": "A. W. Van Buren.",
-        "Judith Andrée-Hanslik": "Judith Andrée-Hanslik.",
-        "Anthony Eric Raubitschek": "A. Raubitschek.",
-        "Hans Georg Gundel": "H. Gundel.",
-    }
-
-    COMPLEX_AUTHORS: dict[str, str] = {
-        "Ernst Hugo Berger": "Berger.",
-        "Adolf Berger": "Berger.",
-        "Ludo Moritz Hartmann": "Hartmann.",
-        "Richard Hartmann": "Hartmann.",
-        "Maria Assunta Nagl": "Nagl.",
-        "Alfred Nagl": "Nagl.",
-        "Alfred Philippson": "Philippson.",
-        "Johannes Schmidt (Epigraphiker)": "J. Schmidt.",
-        "Johannes Schmidt (Philologe)": "J. Schmidt.",
-        "Ernst Schwabe": "Schwabe.",
-        "Ludwig Schwabe": "Schwabe.",
-        "Hans Schaefer": "Hans Schaefer.",
-    }
-
-    REGEX_COMPLEX = re.compile(rf"REAutor\|(?P<author>{'|'.join(set(COMPLEX_AUTHORS.values()))})")
-
-    @classmethod
-    def get_author_mapping(cls) -> dict[str, str]:
-        authors = Authors()
-        author_raw_mapping: dict[str, list[str]] = {}
-        for author in authors.authors_mapping:
-            if isinstance(authors.authors_mapping[author], (dict, list)):
-                continue
-            if authors.authors_mapping[author] not in author_raw_mapping:
-                author_raw_mapping[authors.authors_mapping[author]] = []
-            author_raw_mapping[authors.authors_mapping[author]].append(author)
-        author_mapping = {}
-        for key, value in author_raw_mapping.items():
-            value = [item for item in value if item[-1] == "."]
-            if len(value) == 1 and value[0][-1] == ".":
-                author_mapping[key] = value[0]
-            else:
-                last_name = f"{key.split(" ")[-1]}."
-                name_list = []
-                for name in key.split(" ")[0:-1]:
-                    name_list.append(f"{name[0]}.")
-                name_list.append(last_name)
-                long_last_name = " ".join(name_list)
-                if last_name in value:
-                    author_mapping[key] = last_name
-                elif long_last_name in value:
-                    author_mapping[key] = long_last_name
-
-        author_mapping.update(cls.ADDITIONAL_AUTHORS)
-        author_mapping.update(cls.COMPLEX_AUTHORS)
-        return author_mapping
-
-    @classmethod
-    def adjust_author(cls, input_str: str, mapping: dict[str, str]) -> str:
-        for author in mapping:
-            input_str = re.sub(rf"{{{{REAutor\|{author}}}}}",
-                               f"{{{{REAutor|{mapping[author]}}}}}",
-                               input_str)
-        if cls.REGEX_COMPLEX.search(input_str):
-            article = Article.from_text(input_str.strip())
-            input_str = cls.REGEX_COMPLEX.sub(rf"REAutor|\g<author>|{article["BAND"].value}", input_str)
-        return input_str
 
     @staticmethod
     def adjust_end_column(article_text, register, idx):
